# demo/src/main/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Blog",
    "description": "Api Description",
    "contact": {
      "email": "Petrov.v.163@yandex.ru"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/home"
    }
  ],
  "tags": [
    {
      "name": "post",
      "description": "Access to Petstore orders"
    },
    {
      "name": "MusicGroup",
      "description": "Operations about user",
      "externalDocs": {
        "description": "Find out more about our store",
        "url": "http://swagger.io"
      }
    }
  ],
  "paths": {
    "/api/post": {
      "get": {
        "tags": [
          "post"
        ],
        "summary": "Find post by ID",
        "description": "Returns a single post",
        "operationId": "postGet",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID of post to return",
            "required": true,
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Post"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied",
            "content": {}
          },
          "404": {
            "description": "Post not found",
            "content": {}
          }
        }
      },
      "post": {
        "tags": [
          "post"
        ],
        "summary": "Add new post",
        "operationId": "postAdd",
        "parametes": null,
        "requestBody": {
          "content": {
            "application/RAZOBRATY'SA": {
              "schema": {
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "title of post"
                  },
                  "anons": {
                    "type": "string",
                    "description": "anons of post"
                  },
                  "full_text": {
                    "type": "string",
                    "description": "full text of post"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "405": {
            "description": "Invalid input",
            "content": {}
          }
        }
      },
      "put": {
        "tags": [
          "post"
        ],
        "summary": "Update an existing post",
        "operationId": "postPatch",
        "requestBody": {
          "description": "Post object that needs to be updated",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/Post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid ID supplied",
            "content": {}
          },
          "404": {
            "description": "MusicGroup not found",
            "content": {}
          }
        }
      },
      "delete": {
        "tags": [
          "post"
        ],
        "summary": "Delete a post by id",
        "operationId": "postDelete",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Post id to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied",
            "content": {}
          },
          "404": {
            "description": "MusicGroup not found",
            "content": {}
          }
        }
      }
    },
    "/api/posts": {
      "get": {
        "tags": [
          "post"
        ],
        "summary": "Find all post",
        "description": "Returns all  post",
        "operationId": "postGetAll",
        "parameters": null,
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Posts"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Posts"
                }
              }
            }
          },
          "404": {
            "description": "Posts not found",
            "content": {}
          }
        }
      },
      "delete": {
        "tags": [
          "post"
        ],
        "summary": "Delete all post",
        "operationId": "postDeleteAll",
        "parameters": null,
        "responses": {
          "404": {
            "description": "Posts not found",
            "content": {}
          }
        }
      }
    },
    "/api/group": {
      "get": {
        "tags": [
          "MusicGroup"
        ],
        "summary": "Find MusicGroup by ID",
        "description": "Returns a single MusicGroup",
        "operationId": "groupGet",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID of MusicGroup to return",
            "required": true,
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MusicGroup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicGroup"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied",
            "content": {}
          },
          "404": {
            "description": "Post not found",
            "content": {}
          }
        }
      },
      "post": {
        "tags": [
          "MusicGroup"
        ],
        "summary": "Add new MusicGroup",
        "operationId": "groupAdd",
        "parametes": null,
        "requestBody": {
          "content": {
            "application/RAZOBRATY'SA": {
              "schema": {
                "properties": {
                  "title": {
                    "type": "string",
                    "description": "name of MusicGroup"
                  },
                  "anons": {
                    "type": "string",
                    "description": "style of MusicGroup"
                  },
                  "full_text": {
                    "type": "string",
                    "description": "country of MusicGroup"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "405": {
            "description": "Invalid input",
            "content": {}
          }
        }
      },
      "put": {
        "tags": [
          "MusicGroup"
        ],
        "summary": "Update an existing MusicGroup",
        "operationId": "groupPatch",
        "requestBody": {
          "description": "MusicGroup object that needs to be updated",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MusicGroup"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/MusicGroup"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid ID supplied",
            "content": {}
          },
          "404": {
            "description": "MusicGroup not found",
            "content": {}
          }
        }
      },
      "delete": {
        "tags": [
          "post"
        ],
        "summary": "Delete a MusicGroup by id",
        "operationId": "groupDelete",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "MusicGroup id to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied",
            "content": {}
          },
          "404": {
            "description": "MusicGroup not found",
            "content": {}
          }
        }
      }
    },
    "/api/groups": {
      "get": {
        "tags": [
          "MusicGroup"
        ],
        "summary": "Find all MusicGroups",
        "description": "Returns all MusicGroups",
        "operationId": "groupGetAll",
        "parameters": null,
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/MusicGroup"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MusicGroup"
                }
              }
            }
          },
          "404": {
            "description": "MusicGroups not found",
            "content": {}
          }
        }
      },
      "delete": {
        "tags": [
          "MusicGroup"
        ],
        "summary": "Delete all MusicGroups",
        "operationId": "groupDeleteAll",
        "parameters": null,
        "responses": {
          "404": {
            "description": "MusicGroups not found",
            "content": {}
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Post": {
        "required": [
          "title",
          "anons",
          "full_text"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Post"
          },
          "title": {
            "type": "string",
            "example": "post name"
          },
          "anons": {
            "type": "string",
            "example": "post title"
          },
          "full_text": {
            "type": "string",
            "example": "post full_text"
          },
          "views": {
            "type": "integer",
            "example": 42
          }
        },
        "xml": {
          "name": "Post"
        }
      },
      "MusicGroup": {
        "required": [
          "name",
          "style",
          "country"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "string"
          },
          "category": {
            "$ref": "#/components/schemas/MusicGroup"
          },
          "name": {
            "type": "string",
            "example": "Armant"
          },
          "style": {
            "type": "string",
            "example": "Techno"
          },
          "country": {
            "type": "string",
            "example": "Russia"
          }
        },
        "xml": {
          "name": "MusicGroup"
        }
      },
      "Posts": {
        "required": [
          "name",
          "style",
          "country"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "string"
          },
          "category": {
            "$ref": "#/components/schemas/Posts"
          },
          "title": {
            "type": "string",
            "example": "post name"
          },
          "anons": {
            "type": "string",
            "example": "post title"
          },
          "full_text": {
            "type": "string",
            "example": "post full_text"
          },
          "views": {
            "type": "integer",
            "example": 42
          }
        },
        "xml": {
          "name": "Posts"
        }
      }
    }
  }
}